Separating device definitions
