+++ custom_components/landroid_cloud/device_base.py
@@ -3,14 +3,11 @@ from __future__ import annotations
 
 import logging
 
-from functools import partial
-
 from homeassistant.components.vacuum import (
     ENTITY_ID_FORMAT,
     STATE_DOCKED,
     STATE_ERROR,
     STATE_RETURNING,
-    StateVacuumEntity,
     VacuumEntityFeature,
 )
 from homeassistant.const import CONF_TYPE
@@ -200,49 +197,3 @@ class LandroidCloudBase(Entity):
             device: WorxCloud = self.api.device
             _LOGGER.debug("Sending %s back to dock", self._name)
             await self.hass.async_add_executor_job(device.home)
-
-
-class WorxDevice(LandroidCloudBase, StateVacuumEntity):
-    """Definition of Worx Landroid device."""
-
-    @property
-    def supported_features(self):
-        """Flag which mower robot features that are supported."""
-        return SUPPORT_LANDROID
-
-    async def async_toggle_lock(self):
-        """Toggle locked state."""
-        device: WorxCloud = self.api.device
-        set_lock = not bool(device.locked)
-        _LOGGER.debug("Setting locked state for %s to %s", self._name, set_lock)
-        await self.hass.async_add_executor_job(partial(device.lock, set_lock))
-
-    async def async_toggle_partymode(self):
-        """Toggle partymode state."""
-        device: WorxCloud = self.api.device
-        set_partymode = not bool(device.partymode_enabled)
-        _LOGGER.debug("Setting PartyMode to %s on %s", set_partymode, self._name)
-        await self.hass.async_add_executor_job(
-            partial(device.enable_partymode, set_partymode)
-        )
-
-    async def async_edgecut(self):
-        """Start edgecut routine."""
-        device: WorxCloud = self.api.device
-        _LOGGER.debug("Starting edgecut routine for %s", self._name)
-        await self.hass.async_add_executor_job(device.edgecut)
-
-    async def async_setzone(self, **kwargs):
-        """Set next zone to cut."""
-        device: WorxCloud = self.api.device
-        zone = kwargs["zone"]
-        _LOGGER.debug("Setting zone for %s to %s", self._name, zone)
-        await self.hass.async_add_executor_job(partial(device.setzone, zone))
-
-
-class KressDevice(LandroidCloudBase, StateVacuumEntity):
-    """Definition of Kress device."""
-
-
-class LandxcapeDevice(LandroidCloudBase, StateVacuumEntity):
-    """Definition of Landxcape device."""
