+++ custom_components/landroid_cloud/devices/__init__.py
@@ -1,0 +1,1 @@
+"""Device definitions."""
